requirements.txt
Django==1.8.12
djangorestframework==3.3.2
